+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add examples to swagger documentation
 - Support for config server
 - Add documentation
+- Keycloak security for service-to-service gRPC communication
 
 ## [1.1.0] - 2025-12-03
 ### Added